+++ Src/MoneyFox.Ui/Views/Accounts/AccountModification/AddAccountPage.xaml.cs
@@ -1,4 +1,4 @@
-namespace MoneyFox.Ui.Views.Accounts;
+namespace MoneyFox.Ui.Views.Accounts.AccountModification;
 
 public partial class AddAccountPage
 {
