+++ Src/MoneyFox.Ui/Views/Accounts/AccountModification/AddAccountViewModel.cs
@@ -1,9 +1,9 @@
-namespace MoneyFox.Ui.Views.Accounts;
+namespace MoneyFox.Ui.Views.Accounts.AccountModification;
 
-using Core.Common.Interfaces;
-using Core.Features._Legacy_.Accounts.CreateAccount;
-using Core.Interfaces;
 using MediatR;
+using MoneyFox.Core.Common.Interfaces;
+using MoneyFox.Core.Features._Legacy_.Accounts.CreateAccount;
+using MoneyFox.Core.Interfaces;
 
 internal sealed class AddAccountViewModel : ModifyAccountViewModel
 {
