+++ Src/MoneyFox.Ui/Views/Accounts/AccountModification/ModifyAccountContentView.xaml.cs
@@ -1,4 +1,4 @@
-namespace MoneyFox.Ui.Views.Accounts;
+namespace MoneyFox.Ui.Views.Accounts.AccountModification;
 
 public partial class ModifyAccountContentView : ContentView
 {
